updating schema after merging main

+++ db/schema.rb
@@ -58,7 +58,7 @@ ActiveRecord::Schema[7.0].define(version: 2023_03_16_180414) do
   create_table "samples", force: :cascade do |t|
     t.string "name"
     t.text "description"
-    t.integer "project_id", null: false
+    t.bigint "project_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["name", "project_id"], name: "index_samples_on_name_and_project_id", unique: true
